pyproject.toml
[project]
name = "project-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.1.0",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "gunicorn>=25.1.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-mcp-adapters>=0.1.14",
    "langchain-neo4j>=0.8.0",
    "langchain-openai>=1.1.0",
    "neo4j>=6.1.0",
    "python-dotenv>=1.2.1",
]
